util/cbmem: add type cast
arch_convert_raw_ts_entry returns a uint64_t, which needs to be cast
on ARM systems to avoid a type error.

BUG=b/227871959
TEST=no build errors in downstream

Reviewed-on: https://review.coreboot.org/c/coreboot/+/63317
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ util/cbmem/cbmem.c
@@ -639,7 +639,7 @@ static void print_with_path(struct ts_range_stack *range_stack, const int stackl
 	}
 	if (last_part)
 		printf("%s", last_part);
-	printf(" %lu\n", arch_convert_raw_ts_entry(stamp));
+	printf(" %llu\n", (long long)arch_convert_raw_ts_entry(stamp));
 }
 
 enum timestamps_print_type {
